added form validation

+++ NodeJS_Training-nodeAssignment/myapp/public/javascripts/signInRequest.js
@@ -1,32 +1,46 @@
 
 signIn = (event) => {
-    
+
     event.preventDefault();
     var email = $('#email').val();
     var password = $('#pwd').val();
-   
+
+    if(email.length<1 && password.length<1 )    {
+        
+    confirm("all fields are compulsory");
+    }
+    else {
+        
     $.ajax({
         url: "http://localhost:3000/signin",
         method: "POST",
         data: {
-        
-            email : email,
+
+            email: email,
             password: password
         },
         success: (result) => {
-            console.log("====>",result.success);
-            console.log("====>",result.authorization);
-            localStorage.setItem("token",result.authorization);
-        
-        if(result.success && result.isAdmin){
-            window.location.replace('/homeadmin');
-        }
-        else if(result.success && !result.isAdmin) {
-            window.location.replace('/home');            
-        }
-            
-        } ,
-    
+
+            console.log(result);
+
+            if (result.success) {
+              
+                localStorage.setItem("token", result.authorization);
+
+                if (result.success && result.isAdmin) {
+                    window.location.replace('/homeadmin');
+                }
+                else if (result.success && !result.isAdmin) {
+                    window.location.replace('/home');
+                }
+            }
+            else {
+                alert("Authentication Failed");
+            }
+        },
+
     });
-   
+    }
+
+
 }
